move schemas to folder
